refactor(callbacks): extract Overview panels block to _callbacks_overview.py
Step 7b of the callbacks.py decomposition (issue #87). Second of three
Overview sub-PRs. Builds on PR #104 (overview core) which is stacked
underneath.

## What moves

The Overview panels block — drivers / generation / models leaderboard /
risk / scenarios — moves out of callbacks.py and joins the 5 helpers
from #104 in ``components/_callbacks_overview.py``:

* ``_fetch_generation_cached`` — 3-tier cached EIA-930 fetcher (memory
  → SQLite/API → None). Only consumed by ``_build_generation_panel``,
  so it migrates with the panel rather than to shared.
* ``_build_drivers_panel`` — 3-up KPI cells (Temperature / Wind / Solar)
  with current value + 24h sparkline.
* ``_drivers_empty`` / ``_driver_cell_empty`` — placeholders.
* ``_FUEL_STACK_ORDER`` / ``_FUEL_DISPLAY`` — module-level constants
  defining the stacked-area chart ordering and per-fuel display config.
* ``_build_generation_panel`` — stacked-area fuel mix + 3-up sub-MetricsBar
  (Net Load / Renewable / Largest source).
* ``_generation_empty`` — placeholder.
* ``_build_models_leaderboard`` — 5-up MetricsBar leaderboard with
  hero pick by lowest MAPE.
* ``_build_risk_insight`` — 3-sentence narrative composing demand-anomaly
  + weather-severity clauses.
* ``_build_scenarios_panel`` — heuristic linear-sensitivity scenario
  forecast with KPI bar + baseline-vs-scenario chart.
* ``_driver_sparkline`` — 60px minimal-axis sparkline; only used by
  ``_build_drivers_panel`` so it migrates with the panel.

Also removed a now-orphaned ``# ── Forecast tab helpers (R4a) ──``
section comment that originally labeled the block — its functions are
all Overview helpers in current use, despite the historical name.

## Internal helpers stay internal

``_drivers_empty``, ``_driver_cell_empty``, ``_driver_sparkline``,
``_generation_empty``, ``_FUEL_STACK_ORDER``, ``_FUEL_DISPLAY`` are not
re-exported from callbacks.py — nothing outside this module references
them. The 5 callback-facing builders + ``_fetch_generation_cached``
(tested directly) ARE re-exported via the import shim.

## Public-import surface preserved

callbacks.py re-imports each user-facing function by name.
``register_callbacks`` continues to call ``_build_models_leaderboard``,
``_build_risk_insight``, ``_build_drivers_panel``,
``_build_generation_panel``, ``_build_scenarios_panel`` directly through
the same namespace. Tests using ``from components.callbacks import
_fetch_generation_cached`` continue to resolve.

Added ``# noqa: F401`` markers on ``_EIA_FUEL_MAP`` and
``_fetch_generation_cached`` — these are re-exports for tests that
``ruff --fix`` would otherwise drop as unused.

## Test patches

Existing test patches don't need updating:

* ``_fetch_generation_cached`` tests patch ``config.EIA_API_KEY`` and
  ``data.eia_client.fetch_generation_by_fuel`` (both lazy-imported
  inside the function body, so the patches still hit at call time).
* ``_GENERATION_CACHE`` tests mutate the dict via
  ``cb._GENERATION_CACHE[...]`` — the cache is the same instance whether
  accessed through callbacks.py's namespace or the overview module's.
  Mutations from one perspective are visible from the other.

## Size impact

callbacks.py: 5,369 → 4,686 (–683 lines)
_callbacks_overview.py: 383 → 1,089 (+706 lines, panels block + module
constants + section comment)

**callbacks.py crosses below 5,000 lines for the first time.**

Cumulative progress on the decomposition:
* PR #98 (shared):       7,167 → 7,051 (–116)
* PR #99 (us_grid):      7,051 → 6,389 (–662)
* PR #100 (models):      6,389 → 6,183 (–206)
* PR #101 (alerts):      6,183 → 5,978 (–205)
* PR #102 (generation):  5,978 → 5,802 (–176)
* PR #103 (weather):     5,802 → 5,668 (–134)
* PR #104 (overview 7a): 5,668 → 5,369 (–299)
* this PR (overview 7b): 5,369 → 4,686 (–683)

Total: −2,481 lines from the monolith across 8 PRs (−34.6%).

## Tests

All 1,568 unit tests pass locally. Ruff clean.
